[#49]refactor. 컴포넌트 내부 util폴더 정리
